+++ MbitMoreServiceDAL.h
@@ -40,15 +40,21 @@ public:
 
   void notifySharedData();
 
-  /**
-   * Set value to Slots.
-   */
-  void setSharedData(int index, int value);
+/**
+ * @brief Set value to Shared Data
+ * 
+ * @param index index of the data
+ * @param value value of the data
+ */
+  void setSharedData(int index, float value);
 
   /**
-   * Get value to Slots.
+   * @brief Get value of the Shared Data
+   * 
+   * @param index index of the data
+   * @return float the value of the data
    */
-  int getSharedData(int index);
+  float getSharedData(int index);
 
   /**
    * Callback. Invoked when AnalogIn is read via BLE.
